Added Request and Response.Exceptions.

+++ src/main/java/org/example/repository/OfferRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface OfferRepository extends JpaRepository<Offer,Integer> {
+public interface OfferRepository extends JpaRepository<Offer,Long> {
     @Query(value = "SELECT * FROM touristvoucher.offer", nativeQuery = true)
     public List<Offer> findAllOffer ();
 }
